testbench: Sample rate converter (SRC) setup tool
This change provides a tool to set up the conversions for needed SRC input
and output rates and set the quality parameters for conversions. The generated
coefficients and setup tables can be simply copied to SOF source directory
src/include/reef/audio/coefficients/src for updated conversions and/or
performance. GNU Octave and octave-signal package are needed to run the
scripts.

Signed-off-by: Seppo Ingalsuo <seppo.ingalsuo@linux.intel.com>